src/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-firestore==2.*
openai==1.60.*
langdetect==1.*
googletrans==4.0.2
PyPDF2==3.*